+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Changed
 
 ### Removed
+- [MINOR] removed `LocalStorageClient` code from `useLocalStorageState` and import from core-js.
 
 ## [0.7.0] - 2021-11-21