+++ CHANGELOG.md
@@ -14,7 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - [MAJOR] Changed `Router` and `SubRouter` to require object form routes instead of children
 
 ### Removed
-- [MINOR] Removed `LocalStorageClient` from `useLocalStorageState` (import from core-js)
+- [MINOR] Removed `LocalStorageClient` from `useLocalStorageState` (now import from core-js)
 
 ## [0.7.0] - 2021-11-21