+++ CHANGELOG.md
@@ -10,8 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ### Added
 
 ### Changed
-- [MAJOR] Removed `Route` from routing
 - [MAJOR] Changed `Router` and `SubRouter` to require object form routes instead of children
+- [MAJOR] Removed `Route` from routing
 
 ### Removed
 - [MINOR] Removed `LocalStorageClient` from `useLocalStorageState` (now import from core-js)