+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 ## [Unreleased]
 
 ### Added
+-[MINOR] Added `useRouteParam`, `useStringRouteParam` and `useNumberRouteParam`
 
 ### Changed
 - [MAJOR] Changed `Router` and `SubRouter` to require object form routes instead of children